requirements.txt
altair @ file:///C:/b/abs_0b1avjyzoc/croot/altair_1743016738596/work
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==6.1.0
certifi==2025.7.14
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
Django==4.2.11
Flask==2.3.3
gitdb==4.0.12
GitPython==3.1.45
greenlet @ file:///C:/b/abs_bbl_18a2tm/croot/greenlet_1733860081977/work
idna==3.10
itsdangerous @ file:///C:/b/abs_c4vwgdr5yn/croot/itsdangerous_1716533399914/work
Jinja2==3.1.6
jsonschema @ file:///C:/b/abs_4d8dstds3t/croot/jsonschema_1753206424024/work
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
narwhals==2.0.1
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
Pygments==2.19.2
PyMySQL==1.1.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
SQLAlchemy @ file:///C:/b/abs_aa2izo0xio/croot/sqlalchemy_1718378184063/work
sqlparse==0.5.3
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
